apps/api/pyproject.toml
[project]
name = "waygate-api"
version = "0.1.0"
description = "WayGate API Server"
readme = "README.md"
authors = [
  { name = "Buck Brady", email = "22723438+voidrot@users.noreply.github.com" },
]
requires-python = ">=3.14"
dependencies = [
  "fastapi[standard]>=0.135.3",
  "opentelemetry-instrumentation-fastapi>=0.62b0",
  "pydantic>=2.12.5",
  "waygate-core",
  "waygate-plugin-communication-http",
  "waygate-plugin-communication-nats",
  "waygate-plugin-communication-rq",
  "waygate-plugin-local-storage",
  "waygate-plugin-webhook-agent-session",
  "waygate-plugin-webhook-generic",
  "waygate-webhooks",
]

[dependency-groups]
dev = ["pytest>=9.0.3", "pytest-cov>=7.1.0"]

[project.scripts]
waygate-api = "waygate_api:main"

[build-system]
requires      = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[tool.uv.sources]
waygate-core                         = { workspace = true }
waygate-plugin-communication-http    = { workspace = true }
waygate-plugin-communication-nats    = { workspace = true }
waygate-plugin-communication-rq      = { workspace = true }
waygate-plugin-local-storage         = { workspace = true }
waygate-plugin-webhook-agent-session = { workspace = true }
waygate-plugin-webhook-generic       = { workspace = true }
waygate-webhooks                     = { workspace = true }

[tool.fastapi]
entrypoint = "waygate_api.server:app"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
  "-ra",
  "--strict-config",
  "--strict-markers",
  "--cov=waygate_api",
  "--cov-report=term-missing",
]
